backend/requirements.txt
alembic==1.13.1
annotated-types==0.6.0
anyio==4.2.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgi-lifespan==2.1.0
async-timeout==4.0.3
asyncio==3.4.3
asyncpg==0.29.0
attrs==23.2.0
bcrypt==4.1.2
black==24.2.0
certifi==2024.2.2
cffi==1.16.0
cfgv==3.4.0
click==8.1.7
colorama==0.4.6
coverage==7.4.1
distlib==0.3.8
dnspython==2.5.0
ecdsa==0.18.0
email-validator==2.1.0.post1
exceptiongroup==1.2.0
execnet==2.0.2
fastapi==0.109.2
filelock==3.13.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.3
httpx==0.26.0
identify==2.5.34
idna==3.6
iniconfig==2.0.0
isort==5.13.2
loguru==0.7.2
Mako==1.3.2
MarkupSafe==2.1.5
mypy==1.8.0
mypy-extensions==1.0.0
nodeenv==1.8.0
outcome==1.3.0.post0
packaging==23.2
passlib==1.7.4
pathlib==1.0.1
pathspec==0.12.1
platformdirs==4.2.0
pluggy==1.4.0
pre-commit==3.6.1
pyasn1==0.5.1
pycparser==2.21
pydantic==2.6.1
pydantic-settings==2.2.0
pydantic_core==2.16.2
pytest==8.0.0
pytest-asyncio==0.23.5
pytest-cov==4.1.0
pytest-xdist==3.5.0
python-decouple==3.8
python-dotenv==1.0.1
python-jose==3.3.0
python-slugify==8.0.4
PyYAML==6.0.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.0b3
starlette==0.36.3
text-unidecode==1.3
tomli==2.0.1
trio==0.24.0
typing_extensions==4.9.0
uvicorn==0.27.1
virtualenv==20.25.0
